requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
jinja2==3.1.6
python-dotenv==1.2.1
google-generativeai==0.8.6
chromadb==1.4.0
pypdf==6.5.0
python-multipart==0.0.21
sqlalchemy==2.0.45
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.5.0
email-validator==2.3.0
pydantic==2.12.5
